Updating more Mempool & Manager App APIs with basic testings

+++ Manager/app/main.py
@@ -16,10 +16,10 @@ bucket_name = 'ece1779-winter23-a2-bucket'
 manager_url = 'http://127.0.0.1:5000'
 memcache_pool_url = 'http://127.0.0.1:5002'
 autoscaler_url = 'http://127.0.0.1:5003'
-s3client = None
-rds_client = None
 
 def provision_aws():
+    global s3client
+    global rds_client
     s3client = boto3.client('s3', region_name='us-east-1')
     rds_client = boto3.client('rds', region_name='us-east-1')
     print("Provision done")
@@ -72,6 +72,8 @@ def test_hash():
 #test_bucket()
 provision_aws()
 test_hash()
+print(s3client)
+print(rds_client)
 
 @manager.route('/')
 def main():
@@ -123,21 +125,19 @@ def getFromS3():
     #pending rds getting filename
     filename = request.args.get('name')
     full_file_path = os.path.join(os_file_path, filename)
-
-    s3client.download_file(bucket_name, filename, full_file_path)
-    fileInfo = None
-    if fileInfo == None:
-        response = manager.response_class(
-            response=json.dumps("File Not Found"),
-            status=400,
-            mimetype='application/json'
-        )
-    else:
-        full_file_path = fileInfo.location
+    checkFile = s3client.list_objects_v2(Bucket=bucket_name, Prefix=filename)
+    if "Contents" in checkFile:
+        s3client.download_file(bucket_name, filename, full_file_path)
         value = Path(full_file_path).read_text()
         response = manager.response_class(
             response=json.dumps(value),
             status=200,
+            mimetype='application/json'
+        )
+    else:
+        response = manager.response_class(
+            response=json.dumps("File Not Found"),
+            status=400,
             mimetype='application/json'
         )
 
@@ -163,7 +163,7 @@ def retrieve(key_value):
             }
             response = manager.response_class(
                 response=json.dumps(resp),
-                status=200,
+                status=400,
                 mimetype='application/json'
             )
         else:
@@ -185,7 +185,7 @@ def retrieve(key_value):
         resp["success"] = "true"
         resp["key"] = key_value
         resp["content"] = bytes.decode(content)
-        response = application.response_class(
+        response = manager.response_class(
             response=json.dumps(resp),
             status=200,
             mimetype='application/json'
@@ -199,8 +199,10 @@ def delete_all():
     No inputs required
     """
     response = s3client.list_objects_v2(Bucket=bucket_name)
-    for obj in response['Contents']:
-        s3client.delete_object(Bucket=bucket_name, Key=obj['Key'])
+    #print(response)
+    if 'Contents' in response: 
+        for obj in response['Contents']:
+            s3client.delete_object(Bucket=bucket_name, Key=obj['Key'])
     #pending removal from rds
     requests.post(memcache_pool_url + '/clear')
     resp = {
@@ -217,7 +219,7 @@ def delete_all():
 @manager.route('/clear_memcache', methods=['POST'])
 def clear_memcache():
     """
-    Remove all the key and values (files, images) from the database and filesystem
+    Remove all the key and values (files, images) from the memcache
     No inputs required
     """
     requests.post(memcache_pool_url + '/clear')
@@ -233,7 +235,78 @@ def clear_memcache():
 
 @manager.route('/api/configure_cache', methods=['POST'])
 def configure_memcache():
-    pass
+    requestJson = {}
+    if 'mode' in request.args: 
+        requestJson["mode"] = request.args.get('mode')
+        if requestJson["mode"] != "manual" and requestJson["mode"] != "auto": 
+            resp = {
+                "success": "false",
+                "error": {
+                        "code": 400, 
+                        "message": "Invalid Mode Configuration"
+                        }
+            }               
+            response = manager.response_class(
+                response=json.dumps(resp),
+                status=400,
+                mimetype='application/json'
+            )
+            return response
+    if 'numNodes' in request.args: 
+        requestJson["numNodes"] = int(request.args.get('numNodes'))
+        if requestJson["numNodes"] > 8 and requestJson["numNodes"] < 1: 
+            resp = {
+                "success": "false",
+                "error": {
+                        "code": 400, 
+                        "message": "Number of nodes must be between 1 and 8"
+                        }
+            }               
+            response = manager.response_class(
+                response=json.dumps(resp),
+                status=400,
+                mimetype='application/json'
+            )
+            return response
+    if 'cacheSize' in request.args: 
+        requestJson["cacheSize"] = int(request.args.get('cacheSize'))
+    if 'policy' in request.args: 
+        requestJson["policy"] = request.args.get('policy')
+        if requestJson["policy"] != "LRU" and requestJson["policy"] != "RR": 
+            resp = {
+                "success": "false",
+                "error": {
+                        "code": 400, 
+                        "message": "Invalid Mode Configuration"
+                        }
+            }               
+            response = manager.response_class(
+                response=json.dumps(resp),
+                status=400,
+                mimetype='application/json'
+            )
+            return response
+    if 'expRatio' in request.args: 
+        requestJson["expRatio"] = request.args.get('expRatio')
+    if 'shrinkRatio' in request.args: 
+        requestJson["shrinkRatio"] = request.args.get('shrinkRatio')
+    if 'maxMiss' in request.args: 
+        requestJson["maxMiss"] = request.args.get('maxMiss')
+    if 'minMiss' in request.args: 
+        requestJson["minMiss"] = request.args.get('minMiss')
+    
+    #Pending: if mode is auto specified, need to notify the auto scaler
+
+    requests.post(memcache_pool_url + '/configure', params=requestJson)
+    resp = {"success": "true"}
+    for key in requestJson:
+        resp[key] = requestJson[key]
+    response = manager.response_class(
+        response=json.dumps(resp),
+        status=200,
+        mimetype='application/json'
+    )
+    return response
 
 @manager.route('/api/getNumNodes', methods=['POST'])
 def getNumNodes_Manager():
@@ -252,12 +325,31 @@ def getNumNodes_Manager():
 
 @manager.route('/api/getRate', methods=['POST'])
 def getRate():
+    requestJson = {}
+    if 'rate' in request.args: 
+        requestJson["rate"] = request.args.get('rate')
+        if requestJson["rate"] != "miss" and requestJson["rate"] != "hit": 
+            resp = {
+                "success": "false",
+                "error": {
+                        "code": 400, 
+                        "message": "Only miss and hit rates are available"
+                        }
+            }               
+            response = manager.response_class(
+                response=json.dumps(resp),
+                status=400,
+                mimetype='application/json'
+            )
+            return response
     pass
 
 @manager.route('/api/list_keys', methods=['POST'])
 def list_keys():
     res = requests.post(memcache_pool_url + '/all_keys')
-    all_keys = res.content
+    json_response = res.json()
+    print(json_response)
+    all_keys = json_response["all_keys"]
     resp = {
         "success" : "true", 
         "keys": all_keys
